// src/app/model/partnerMasterVO.model.ts
export class PartnerMasterVO {
    partnerId?: number;
    partnerName?: string;
    bpNo?: string;
    address?: string;
    city?: string;
    state?: string;
    pincode?: string;
    businessBackground?: string;
    partnerType?: string;
    experienceYears?: number;
    region?: string;
    mobileNo?: string;
    emailId?: string;
    aadhaarNo?: string;
    panNo?: string;
    drivingLicense?: string;
    companyName?: string;
    companyAddress?: string;
    companyCity?: string;
    companyState?: string;
    companyPincode?: string;
    companyMobileNo?: string;
    companyEmailId?: string;
    isGst?: boolean;
    gstNo?: string;
    isAddressProof?: boolean;
    managerName?: string;
    managerMobileNo?: string;
    managerEmailId?: string;
    isDocumentsSubmitted?: boolean;
    createdBy?: string;
    creationDate?: Date;
    updatedBy?: string;
    updationDate?: Date;
  }
  