// src/app/model/vehicleDetailsVO.model.ts
export class VehicleDetailsVO {
    vehicleId?: number;
    vehicleCode?: string;
    vTypeId?: number;
    model?: string;
    vehicleType?: string;
    seatingCapacity?: number;
    licensePlateNo?: string;
    rcNumber?: string;
    insuranceNo?: string;
    pucNo?: string;
    isDocumentsSubmitted?: boolean;
    partnerId?: number;
    bpCode?: string;
    partnerName?: string;
    isApproved?: boolean;
    approverName?: string;
    createdBy?: string;
    creationDate?: Date;
    updatedBy?: string;
    updationDate?: Date;

}